+++ python/sdss_access/sync/http.py
@@ -19,8 +19,8 @@ class HttpAccess(SDSSPath):
     """Class for providing HTTP access via urllib.request (python3) or urllib2 (python2) to SDSS SAS Paths
     """
 
-    def __init__(self, verbose=False):
-        super(HttpAccess, self).__init__()
+    def __init__(self, verbose=False, public=False, release=None):
+        super(HttpAccess, self).__init__(public=public, release=release, verbose=verbose)
         self.verbose = verbose
         self._remote = False